@@ -82,8 +82,8 @@ dependencies:
   # -- localization -- #
   flutter_localizations:
     sdk: flutter
-  slang: ^3.31.2
-  slang_flutter: ^3.31.0
+  slang: ^4.4.1
+  slang_flutter: ^4.4.0
 
   # -- ui -- #
   flex_color_scheme: ^7.3.1
@@ -117,6 +117,10 @@ dependencies:
   file_picker: ^8.1.2
   flutter_dash: ^1.0.0
 
+  skit:
+    git:
+      url: https://github.com/nikondrat/skit.git
+
 dev_dependencies:
   # -- testing -- #
   flutter_test:
@@ -130,7 +134,7 @@ dev_dependencies:
   # -- codegen -- #
   build_runner: ^2.4.11
   mobx_codegen: ^2.6.0+1
-  slang_build_runner: ^3.31.0
+  slang_build_runner: ^4.4.2
 
 flutter:
   uses-material-design: true
