@@ -1,5 +1,5 @@
-export 'config.dart';
 export 'colors.dart';
 export 'endpoint.dart';
 export 'text_styles.dart';
 export 'generated/generated.dart';
+export 'config.dart';
