@@ -3,6 +3,7 @@ import 'package:go_router/go_router.dart';
 import 'package:mama/src/data.dart';
 import 'package:provider/provider.dart';
 import 'package:reactive_forms/reactive_forms.dart';
+import 'package:skit/skit.dart';
 
 class RegisterFillName extends StatelessWidget {
   const RegisterFillName({
@@ -61,7 +62,8 @@ class RegisterFillName extends StatelessWidget {
                               // const EdgeInsets.symmetric(horizontal: 16),
                               title: t.register.next,
                               isSmall: false,
-                              padding: const EdgeInsets.symmetric(horizontal: 10),
+                              padding:
+                                  const EdgeInsets.symmetric(horizontal: 10),
                               textStyle: textTheme.bodyMedium?.copyWith(
                                   color: isValid
                                       ? AppColors.primaryColor
