@@ -1,19 +1,20 @@
 import 'package:mama/src/data.dart';
 import 'package:mobx/mobx.dart';
+import 'package:skit/skit.dart';
 
 part 'groups.g.dart';
 
 class GroupsStore extends _GroupsStore with _$GroupsStore {
   GroupsStore({
     required super.fetchFunction,
-    required super.restClient,
+    required super.apiClient,
   });
 }
 
 abstract class _GroupsStore extends PaginatedListStore<GroupItem> with Store {
   _GroupsStore({
     required super.fetchFunction,
-    required super.restClient,
+    required super.apiClient,
   }) : super(
             basePath: Endpoint().groups,
             transformer: (raw) {
