@@ -1,8 +1,9 @@
 import 'package:mama/src/data.dart';
+import 'package:skit/skit.dart';
 
 class SchoolsStore extends PaginatedListStore {
   SchoolsStore({
-    required super.restClient,
+    required super.apiClient,
     required super.fetchFunction,
   }) : super(
           basePath: Endpoint().schools,
