@@ -1,9 +1,10 @@
 import 'package:mama/src/data.dart';
+import 'package:skit/skit.dart';
 
 class SchoolStore extends SingleDataStore<SchoolModel> {
-  SchoolStore({required RestClient restClient})
+  SchoolStore({required ApiClient apiClient})
       : super(
-          fetchFunction: (_) => restClient.get(Endpoint.school),
+          fetchFunction: (_) => apiClient.get(Endpoint.school),
           transformer: (v) {
             final data = v?['online_school'];
             return SchoolModel.fromJson(data);
